services/deduplication/requirements.txt
kafka-python==2.0.2
redis==5.0.1
datasketch==1.6.4
